Some small test framework clean up
