+++ tests/testing-framework/src/runtimes/in_process_spin.rs
@@ -2,7 +2,7 @@
 
 use crate::{
     http::{Request, Response},
-    Runtime,
+    Runtime, ServicesConfig, TestEnvironment, TestEnvironmentConfig,
 };
 use anyhow::Context as _;
 
@@ -54,5 +54,50 @@ impl InProcessSpin {
 impl Runtime for InProcessSpin {
     fn error(&mut self) -> anyhow::Result<()> {
         Ok(())
+    }
+}
+
+impl TestEnvironmentConfig<InProcessSpin> {
+    pub fn in_process(
+        services_config: ServicesConfig,
+        preboot: impl FnOnce(&mut TestEnvironment<InProcessSpin>) -> anyhow::Result<()> + 'static,
+    ) -> Self {
+        Self {
+            services_config,
+            create_runtime: Box::new(|env| {
+                preboot(env)?;
+                tokio::runtime::Runtime::new()
+                    .context("failed to start tokio runtime")?
+                    .block_on(async {
+                        use spin_trigger::{
+                            loader::TriggerLoader, HostComponentInitData, RuntimeConfig,
+                            TriggerExecutorBuilder,
+                        };
+                        use spin_trigger_http::HttpTrigger;
+                        let locked_app = spin_loader::from_file(
+                            env.path().join("spin.toml"),
+                            spin_loader::FilesMountStrategy::Direct,
+                            None,
+                        )
+                        .await?;
+                        let json = locked_app.to_json()?;
+                        std::fs::write(env.path().join("locked.json"), json)?;
+
+                        let loader = TriggerLoader::new(env.path().join(".working_dir"), false);
+                        let mut builder = TriggerExecutorBuilder::<HttpTrigger>::new(loader);
+                        // TODO(rylev): see if we can reuse the builder from spin_trigger instead of duplicating it here
+                        builder.hooks(spin_trigger::network::Network::default());
+                        let trigger = builder
+                            .build(
+                                format!("file:{}", env.path().join("locked.json").display()),
+                                RuntimeConfig::default(),
+                                HostComponentInitData::default(),
+                            )
+                            .await?;
+
+                        Result::<_, anyhow::Error>::Ok(InProcessSpin::new(trigger))
+                    })
+            }),
+        }
     }
 }
